requirements.txt
# NaturalLab: Multi-Modal Tracking and Analysis System
# Core dependencies

# Core scientific computing
numpy>=1.19.0
pandas>=1.1.0
scipy>=1.5.0

# Computer Vision
opencv-python>=4.5.0
pillow>=10.0.0

# Deep Learning Framework
torch>=1.7.0
torchvision>=0.8.0

# Detection Models
ultralytics>=8.0.0  # YOLOv8/YOLOv11
transformers>=4.46.0  # OWLv2, CLIP, VLMs

# Tracking
deep-sort-realtime>=1.3.0
filterpy>=1.4.5  # Kalman filtering

# Pose Estimation (optional)
mediapipe>=0.8.0

# Vision-Language Models
open-clip-torch>=2.0.0

# Data formats
h5py>=3.0.0
pyyaml>=5.4.0
pyxdf>=1.16.0  # XDF file handling

# Lab Streaming Layer (Acquisition)
pylsl>=1.16.0

# Visualization
matplotlib>=3.3.0
seaborn>=0.11.0
tqdm>=4.50.0

# Metrics
torchmetrics>=0.7.0

# Utilities
gdown>=4.4.0  # For model downloads
requests>=2.25.0
